Move filename parsing functions to their own module
