+++ src/views/dataset/NewDataset.vue
@@ -51,6 +51,7 @@ import store from "@/store";
 import { IGirderSelectAble } from "@/girder";
 import GirderLocationChooser from "@/components/GirderLocationChooser.vue";
 import { IDataset } from "@/store/model";
+import { collectFilenameMetadata } from "@/utils/parsing";
 
 interface FileUpload {
   file: File;
@@ -96,32 +97,6 @@ function findCommonPrefix(strings: string[]): string {
   }
 
   return result.join("");
-}
-
-function parseTime(s: string): string | null {
-  const re = /time(\d+)/;
-  const match = s.match(re);
-
-  return match ? match[1] : null;
-}
-
-function parseXY(s: string): string | null {
-  const re = /XY(\d+)/;
-  const match = s.match(re);
-
-  return match ? match[1] : null;
-}
-
-function collectFilenameMetadata(
-  filenames: string[]
-): { xy: string[]; t: string[] } {
-  const times = filenames.map(parseTime);
-  const xys = filenames.map(parseXY);
-
-  return {
-    t: times.filter(d => d !== null).sort(),
-    xy: xys.filter(d => d !== null).sort()
-  };
 }
 
 @Component({
